Create Plus_One.py
Intuition

Let us identify the rightmost digit which is not equal to nine and increase that digit by one. All the following consecutive digits of nine should be set to zero.

Here is the simplest use case which works fine.

simple

Here is a slightly complicated case which still passes.

more

And here is the case which breaks everything, because all the digits are nines.

handle

In this case, we need to set all nines to zero and append 1 to the left side of the array.

append

Algorithm

Move along the input array starting from the end of array.

Set all the nines at the end of array to zero.

If we meet a not-nine digit, we would increase it by one. The job is done - return digits.

We're here because all the digits were equal to nine. Now they have all been set to zero. We then append the digit 1 in front of the other digits and return the result.

Implementation